switch editor and viewer
