+++ app/(playground)/p/[agentId]/beta-proto/viewer.tsx
@@ -1,0 +1,18 @@
+import bg from "./bg.png";
+import { Header } from "./header";
+
+export function Viewer() {
+	return (
+		<div
+			className="w-full h-screen bg-black-100"
+			style={{
+				backgroundImage: `url(${bg.src})`,
+				backgroundPositionX: "center",
+				backgroundPositionY: "center",
+				backgroundSize: "cover",
+			}}
+		>
+			<Header />
+		</div>
+	);
+}
